fopix to javix: fix unbound variable
Problem:

When a variable is defined, it was not bound to its value in the environment.
Consequently, when I try to get the value associated to this variable, I got
a None value, and the compiler sent me an exception.

Solution:

When I compile a DefVal (fopix), I just bind this variable to its value
in the environment.

--