fopix to javix: S.DefVal + S.Var done (-.-)//
Ok let me explain what I did.

On the stack, when you put integers, you put integer values (not boxed values)

**Why?**

When you want to apply a binary operation (iadd, imul, ...)
it only works with integer values, not boxed values.
So, it is normal to not box/unbox integer values.
BUT, you may want to do this if you need to store your value in a variable.

If you need to put an integer into a variable, then you must box it.
That is to say that a variable cannot store a value, but an address.

**The problem of DefVar and Var**

That is why when you compile DefVal, the value (on the stack)
bound to the identifier must be stored in a variable.

So, in Javix, you box the value on the top of the stack, and store it.
(the boxed value is removed from the stack and put in the indexed variable).

If you need to get the value of a variable to compile a fopix expression,
in Javix, that means that you must load the value from the variable indexed
by *nextvar*, push it on the stack, and unbox it.

Take a took on javix.fx, launch ```make jflap```, and see the result.

--