fopix to javix : Let done (need to be checked)
Let is done

**Known bug**

Every javix programs must end with *ireturn*.
I integrated that, but it does not work because there is
no integer in the stack.

Even if I compile and run:

```val x = 1```

In javix, 1 is stored in x, so it is not in the stack,
but I *ireturn* from an ampty stack.

FIX ME

--