fopix to javix: what I did so far
```
 a = b
 a < b
 ...
```
are compiled as:

```
 if a = b then true else false
 if a < b then true else false
```

because I don't have a specific javix operation to compare values.
The only javix instruction I can use is If_icmp.

I tried to compile those comparison but they only work
if the condition is false. Still, the compilation is incomplete.

Here is the structure of If_icmp (according to what I think it should be):

```
... // some code
if_icmp <cond> <ltrue>  // ltrue is the label I must jump to if <cond> is true
|                       // I think I should generate this label
|
| <code_f>              // code to execute if the condition is false
|
| goto < endif_<n> >    // n in [0, +inf[ , I think I should generate this label
|
<ltrue>:
|
| <code_t>              // code to execute if the condition is true
|
< endif_<n> >
|
| ... // next instructions
|

```

If I am not wrong, I should do it for comparison and IFThenElse fopix instruction.
But I have some difficulties to put the labels in the dispatch, so
the javix interpreter can get information on how/wher to jump excatly.

--